localbase-provider/requirements.txt
# Core dependencies
pyyaml>=6.0
docker>=6.0.0
requests>=2.28.0
pydantic>=1.10.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.3.0

# Hardware detection
psutil>=5.9.0
py-cpuinfo>=9.0.0
gputil>=1.4.0
nvidia-ml-py>=11.525.0

# Blockchain client
cosmpy>=0.9.0
bip32>=3.3
mnemonic>=0.20

# Model management
huggingface-hub>=0.14.0
transformers>=4.28.0
torch>=2.0.0
onnx>=1.14.0
onnxruntime>=1.14.0
onnxruntime-gpu>=1.14.0

# Security
cryptography>=40.0.0
